src/McpOrchestrator/requirements.txt
fastmcp>=2.3.0
httpx>=0.27.0
azure-functions>=1.21.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
respx>=0.21.0
